Convert inner classes to top-level classes
