+++ junit-jupiter-params/src/main/java/org/junit/jupiter/params/ParameterizedTestInvocationContext.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2015-2017 the original author or authors.
+ *
+ * All rights reserved. This program and the accompanying materials are
+ * made available under the terms of the Eclipse Public License v1.0 which
+ * accompanies this distribution and is available at
+ *
+ * http://www.eclipse.org/legal/epl-v10.html
+ */
+
+package org.junit.jupiter.params;
+
+import static java.util.Collections.singletonList;
+
+import java.util.List;
+
+import org.junit.jupiter.api.extension.Extension;
+import org.junit.jupiter.api.extension.TestTemplateInvocationContext;
+
+class ParameterizedTestInvocationContext implements TestTemplateInvocationContext {
+
+	private final ParameterizedTestNameFormatter formatter;
+	private final Object[] arguments;
+
+	ParameterizedTestInvocationContext(ParameterizedTestNameFormatter formatter, Object[] arguments) {
+		this.formatter = formatter;
+		this.arguments = arguments;
+	}
+
+	@Override
+	public String getDisplayName(int invocationIndex) {
+		return formatter.format(invocationIndex, arguments);
+	}
+
+	@Override
+	public List<Extension> getAdditionalExtensions() {
+		return singletonList(new ParameterizedTestParameterResolver(arguments));
+	}
+}
